fix: ensure database exists before operation

+++ server/main.go
@@ -57,6 +57,14 @@ type Server struct {
 func (s *Server) Create(ctx context.Context, alb *proto.Album) (*proto.Identifier, error) {
 	opsStarted.Inc()
 
+	if s.db == nil {
+		opsFailed.Inc()
+		log.Println("Unable to connect to database")
+		return nil, status.Error(
+			codes.NotFound, "Unable to connect to database",
+		)
+	}
+
 	result, err := s.db.Exec("INSERT INTO album (title, artist, score, cover) VALUES (?, ?, ?, ?)", alb.Title, alb.Artist, alb.Score, alb.Cover)
 	if err != nil {
 		opsFailed.Inc()
@@ -82,6 +90,14 @@ func (s *Server) Create(ctx context.Context, alb *proto.Album) (*proto.Identifie
 // Opens stream for a streaming read of every album in the database
 func (s *Server) Read(_ *proto.Nil, stream proto.Albums_ReadServer) error {
 	opsStarted.Inc()
+
+	if s.db == nil {
+		opsFailed.Inc()
+		log.Println("Unable to connect to database")
+		return status.Error(
+			codes.NotFound, "Unable to connect to database",
+		)
+	}
 
 	rows, err := s.db.Query("SELECT * FROM album")
 	if err != nil {
@@ -134,6 +150,14 @@ func (s *Server) Read(_ *proto.Nil, stream proto.Albums_ReadServer) error {
 func (s *Server) Update(ctx context.Context, in *proto.UpdateRequest) (*proto.Nil, error) {
 	opsStarted.Inc()
 
+	if s.db == nil {
+		opsFailed.Inc()
+		log.Println("Unable to connect to database")
+		return nil, status.Error(
+			codes.NotFound, "Unable to connect to database",
+		)
+	}
+
 	_, err := s.db.Exec("UPDATE album SET title=?, artist=?, score=?, cover=? WHERE id=?", in.NewAlbum.Title, in.NewAlbum.Artist, in.NewAlbum.Score, in.NewAlbum.Cover, in.OldAlbum.Id)
 	if err != nil {
 		opsFailed.Inc()
@@ -152,6 +176,14 @@ func (s *Server) Update(ctx context.Context, in *proto.UpdateRequest) (*proto.Ni
 func (s *Server) Delete(ctx context.Context, alb *proto.Album) (*proto.Nil, error) {
 	opsStarted.Inc()
 
+	if s.db == nil {
+		opsFailed.Inc()
+		log.Println("Unable to connect to database")
+		return nil, status.Error(
+			codes.NotFound, "Unable to connect to database",
+		)
+	}
+
 	_, err := s.db.Exec("DELETE FROM album WHERE id=?", alb.Id)
 	if err != nil {
 		opsFailed.Inc()
@@ -168,6 +200,14 @@ func (s *Server) Delete(ctx context.Context, alb *proto.Album) (*proto.Nil, erro
 
 func (s *Server) Increment(ctx context.Context, in *proto.Identifier) (*proto.Score, error) {
 	opsStarted.Inc()
+
+	if s.db == nil {
+		opsFailed.Inc()
+		log.Println("Unable to connect to database")
+		return nil, status.Error(
+			codes.NotFound, "Unable to connect to database",
+		)
+	}
 
 	_, err := s.db.Exec("UPDATE album SET score = score + 1 WHERE id=?", in.Id)
 	if err != nil {
@@ -196,6 +236,14 @@ func (s *Server) Increment(ctx context.Context, in *proto.Identifier) (*proto.Sc
 
 func (s *Server) Decrement(ctx context.Context, in *proto.Identifier) (*proto.Score, error) {
 	opsStarted.Inc()
+
+	if s.db == nil {
+		opsFailed.Inc()
+		log.Println("Unable to connect to database")
+		return nil, status.Error(
+			codes.NotFound, "Unable to connect to database",
+		)
+	}
 
 	_, err := s.db.Exec("UPDATE album SET score = score - 1 WHERE id=?", in.Id)
 	if err != nil {
